requirements.txt
colorlog
jupyter
matplotlib
numpy
pandas
plotly
PyYAML
scipy
setuptools